+++ apps/modernization-ui/src/design-system/message/index.ts
@@ -1,1 +1,2 @@
 export { Message } from './Message';
+export { AlertMessage } from './alert-message/AlertMessage';
